drivers: fix issues with create_mutexcond_bulk
Fix possible free-after-free on resource if an error happened.
Use explicit width for condMask.
Refactor code and formatting to meet current guidelines.
Apply clang-format.

YT: RTOS-1425

+++ drivers/res-create.h
@@ -12,11 +12,11 @@
 #define NETLIB_RES_CREATE_H_
 
 #include <stdint.h>
+#include <stddef.h>
 #include <sys/threads.h>
 
 
-__attribute__((nonnull(1)))
-int create_mutexcond_bulk(handle_t *out, size_t n, size_t cond_mask);
+__attribute__((nonnull(1))) int create_mutexcond_bulk(handle_t *out, size_t n, uint32_t cond_mask);
 
 
 #endif /* NETLIB_RES_CREATE_H_ */
